move funcs from utils.py to ParametersBase
