+++ taxcalc/parameters_base.py
@@ -1,6 +1,6 @@
 import os
 import json
-from .utils import expand_array
+import numpy as np
 
 
 class ParametersBase(object):
@@ -28,11 +28,11 @@ class ParametersBase(object):
             cpi_inflated = data.get('cpi_inflated', False)
             values = data['value']
             setattr(self, name,
-                    expand_array(values, inflate=cpi_inflated,
-                                 inflation_rates=getattr(self,
-                                                         '_inflation_rates',
-                                                         None),
-                                 num_years=self._num_years))
+                    self.expand_array(values, inflate=cpi_inflated,
+                                      inflation_rates=getattr(self,
+                                                              '_inflation_rates',
+                                                              None),
+                                      num_years=self._num_years))
         self.set_year(self._start_year)
 
     @property
@@ -223,9 +223,164 @@ class ParametersBase(object):
             cval = getattr(self, name, None)
             if cval is None:
                 continue
-            nval = expand_array(values,
-                                inflate=cpi_inflated,
-                                inflation_rates=inf_rates,
-                                num_years=num_years_to_expand)
+            nval = self.expand_array(values,
+                                     inflate=cpi_inflated,
+                                     inflation_rates=inf_rates,
+                                     num_years=num_years_to_expand)
             cval[(self.current_year - self.start_year):] = nval
         self.set_year(self._current_year)
+
+    @staticmethod
+    def expand_1D(x, inflate, inflation_rates, num_years):
+        """
+        Expand the given data to account for the given number of budget years.
+        If necessary, pad out additional years by increasing the last given
+        year at the provided inflation rate.
+        """
+
+        if isinstance(x, np.ndarray):
+            if len(x) >= num_years:
+                return x
+            else:
+                ans = np.zeros(num_years, dtype='f8')
+                ans[:len(x)] = x
+                if inflate:
+                    extra = []
+                    cur = x[-1]
+                    for i in range(0, num_years - len(x)):
+                        inf_idx = i + len(x) - 1
+                        cur *= (1. + inflation_rates[inf_idx])
+                        extra.append(cur)
+                else:
+                    extra = [float(x[-1]) for i in
+                             range(1, num_years - len(x) + 1)]
+
+                ans[len(x):] = extra
+                return ans.astype(x.dtype, casting='unsafe')
+
+        return ParametersBase.expand_1D(np.array([x]), inflate, inflation_rates, num_years)
+
+    @staticmethod
+    def expand_2D(x, inflate, inflation_rates, num_years):
+        """
+        Expand the given data to account for the given number of budget years.
+        For 2D arrays, we expand out the number of rows until we have num_years
+        number of rows. For each expanded row, we inflate by the given inflation
+        rate.
+        """
+
+        if isinstance(x, np.ndarray):
+
+            # Look for -1s and create masks if present
+            last_good_row = -1
+            keep_user_data_mask = []
+            keep_calc_data_mask = []
+            has_nones = False
+            for row in x:
+                keep_user_data_mask.append([1 if i != -1 else 0 for i in row])
+                keep_calc_data_mask.append([0 if i != -1 else 1 for i in row])
+                if not np.any(row == -1):
+                    last_good_row += 1
+                else:
+                    has_nones = True
+
+            if x.shape[0] >= num_years and not has_nones:
+                return x
+            else:
+
+                if has_nones:
+                    c = x[:last_good_row + 1]
+                    keep_user_data_mask = np.array(keep_user_data_mask)
+                    keep_calc_data_mask = np.array(keep_calc_data_mask)
+
+                else:
+                    c = x
+
+                ans = np.zeros((num_years, c.shape[1]))
+                ans[:len(c), :] = c
+                if inflate:
+                    extra = []
+                    cur = c[-1]
+                    for i in range(0, num_years - len(c)):
+                        inf_idx = i + len(c) - 1
+                        cur = np.array(cur * (1. + inflation_rates[inf_idx]))
+                        extra.append(cur)
+                else:
+                    extra = [c[-1, :] for i in
+                             range(1, num_years - len(c) + 1)]
+
+                ans[len(c):, :] = extra
+
+                if has_nones:
+                    # Use masks to "mask in" provided data and "mask out"
+                    # data we don't need (produced in rows with a None value)
+                    ans = ans * keep_calc_data_mask
+                    user_vals = x * keep_user_data_mask
+                    ans = ans + user_vals
+
+                return ans.astype(c.dtype, casting='unsafe')
+
+        return ParametersBase.expand_2D(np.array(x), inflate, inflation_rates, num_years)
+
+    @staticmethod
+    def strip_Nones(x):
+        """
+        Takes a list of scalar values or a list of lists.
+        If it is a list of scalar values, when None is encountered, we
+        return everything encountered before. If a list of lists, we
+        replace None with -1 and return
+
+        Parameters
+        ----------
+        x: list
+
+        Returns
+        -------
+        list
+        """
+        accum = []
+        for val in x:
+            if val is None:
+                return accum
+            if not isinstance(val, list):
+                accum.append(val)
+            else:
+                for i, v in enumerate(val):
+                    if v is None:
+                        val[i] = -1
+                accum.append(val)
+
+        return accum
+
+    @staticmethod
+    def expand_array(x, inflate, inflation_rates, num_years):
+        """
+        Dispatch to either expand_1D or expand2D depending on the dimension of x
+
+        Parameters
+        ----------
+        x : value to expand
+
+        inflate: Boolean
+            As we expand, inflate values if this is True, otherwise, just copy
+
+        inflation_rate: float
+            Yearly inflation reate
+
+        num_years: int
+            Number of budget years to expand
+
+        Returns
+        -------
+        expanded numpy array
+        """
+        x = np.array(ParametersBase.strip_Nones(x))
+        try:
+            if len(x.shape) == 1:
+                return ParametersBase.expand_1D(x, inflate, inflation_rates, num_years)
+            elif len(x.shape) == 2:
+                return ParametersBase.expand_2D(x, inflate, inflation_rates, num_years)
+            else:
+                raise ValueError("Need a 1D or 2D array")
+        except AttributeError:
+            raise ValueError("Must pass a numpy array")
